+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.2.4] - 2024-09-19
+
 ## Added
 
 - Extended `ClickhouseAPI` pandas table registration to support float columns [#24](https://github.com/ADBond/splinkclickhouse/pull/24)
@@ -58,7 +60,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Basic working version of package with api for `chdb`
 
-[unreleased]: https://github.com/ADBond/splinkclickhouse/compare/v0.2.3...HEAD
+[unreleased]: https://github.com/ADBond/splinkclickhouse/compare/v0.2.4...HEAD
+[0.2.4]: https://github.com/ADBond/splinkclickhouse/compare/v0.2.3...v0.2.4
 [0.2.3]: https://github.com/ADBond/splinkclickhouse/compare/v0.2.2...v0.2.3
 [0.2.2]: https://github.com/ADBond/splinkclickhouse/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/ADBond/splinkclickhouse/compare/v0.2.0...v0.2.1